tick: Remove continous timer mode misbehaviour on MR write
When the continous mode is set and the current timer value
matches what is in timer period (TP), the timer is disabled.
This does not match the behaviour that is described in the arch
specification, nor does any of the major implementations behave
like this.

This fixes that and add a test case testing for it.

* ChangeLog:

2013-07-28  Stefan Kristiansson  <stefan.kristiansson@saunalahti.fi>

	* tick/tick.c (spr_write_ttmr): Do not disable counter when in
	continous mode and timer value mathes the timer period.

* testsuite/test-code-or1k/ChangeLog:

2013-07-28  Stefan Kristiansson  <stefan.kristiansson@saunalahti.fi>

	* tick.c: Add test for continous mode timer start when timer
	period matches timer counter.