cuc: Make cuc run again
This fixes a few issues:
  1. There were un-inited memory issues cause crashes
  2. Make some functions static as they are file private
  3. Add some comments and fix variable names to make code more clear

The main issue was that, and still is, that cuc when detecting the end
address of a function in extract_function, it does not detect code after
any l.jr instructions.  For now I just changed it to detect this case
and skip these functions, rather than crash.