rsp-server: Handle vMustReplyEmpty packets
GDB sends vMustReplyEmpty as a feature test to detect old/broken GDB
server behavior[1]. The correct reply is an empty response, which is
what we already send, but we warn about it. Since we should expect to
receive this packet, silence the warning, and add a comment to ensure
we maintain the expected behavior.

[1]: https://sourceware.org/gdb/onlinedocs/gdb/Packets.html

Signed-off-by: Samuel Holland <samuel@sholland.org>