Abstract the target endianness with a macro
Only a few changes are needed to swap the target instruction endianness.
Most of these places are already conditional on the host endianness, but
the changes are not completely overlapping. Add a macro for the target
endianness so it can be swapped out.

In the future, this choice could come from the config file, but that
requires more refactoring.

Signed-off-by: Samuel Holland <samuel@sholland.org>