or32: Fix overflow detection on -O2 compilation
The check for overflow assumes that the compiler will allow us to detect
when a value overflows to a negative number.  When running GCC with -O2
and above this overflow cannot be detected.  The C standard explains
that overflow behavior is undefined meaning this expected overflow
causing a negative number cannot be guaranteed and it is not.

Multiple overflow flag tests were failing.

Fix the detection buy just using the compiler builtins for detecting
overflow conditions.