or1ksim_trace: Fix warnings in trace api for or1k_trace fusesoc core
This patch exposes some of the instruction trace api's as a fusesoc
core.  This is used for verilator targets to support OpenRISC
instruction trace logging.

This required a bit of cleanup as fusesoc compiles all verilator files
as C++.  The C code in or32.c needed to be updated to use const char *
and other modern, safe, practices.